Initial commit
This introduces the very basic setup that'll be used for interacting with environment variables, local secret overrides using a `.env` file, and a few other bits and bobs.  It also allows for setup and development inside of a VSCode dev container for ease of setup and use.  This initial commit just loads a personal access token and requests repo information from the Github API.