Fetch multiple pages of repos from Github
This is a simple algorithm.  First, get pages starting at index 0.  If we have an error, record it and then bail.  If we get an error response from the api, set it and bail.  Finally, if we have responses, append it to a master list.  This may bloat memory (we could readily iterate over what we've got before we fetch more), but it's a simplistic implementation.